move filter status and sort options to isolated constants file
