+++ web/src/components/events/EventFilters.tsx
@@ -1,20 +1,6 @@
 import React from 'react';
 import { Input, Select } from '@/components/ui';
-
-export const STATUS_OPTIONS = [
-  { value: '', label: 'All statuses' },
-  { value: 'published', label: 'Published' },
-  { value: 'draft', label: 'Draft' },
-  { value: 'draft_on_review', label: 'In Review' },
-  { value: 'cancelled', label: 'Cancelled' },
-  { value: 'archived', label: 'Archived' },
-];
-
-export const SORT_OPTIONS = [
-  { value: 'created_at', label: 'Newest first' },
-  { value: 'start_date', label: 'Start date' },
-  { value: 'title', label: 'Title A–Z' },
-];
+import { STATUS_OPTIONS, SORT_OPTIONS } from '@/constants/event.constants';
 
 interface EventFiltersProps {
   search: string;
