+++ web/src/constants/event.constants.ts
@@ -1,0 +1,14 @@
+export const STATUS_OPTIONS = [
+  { value: '', label: 'All statuses' },
+  { value: 'published', label: 'Published' },
+  { value: 'draft', label: 'Draft' },
+  { value: 'draft_on_review', label: 'In Review' },
+  { value: 'cancelled', label: 'Cancelled' },
+  { value: 'archived', label: 'Archived' },
+];
+
+export const SORT_OPTIONS = [
+  { value: 'created_at', label: 'Newest first' },
+  { value: 'start_date', label: 'Start date' },
+  { value: 'title', label: 'Title A–Z' },
+];
